Extract packages from util
