+++ pkg/authentication/scram/scram_enabler_test.go
@@ -3,10 +3,9 @@ package scram
 import (
 	"testing"
 
-	"github.com/mongodb/mongodb-kubernetes-operator/pkg/util/authtypes"
-	"github.com/mongodb/mongodb-kubernetes-operator/pkg/util/constants"
-
+	"github.com/mongodb/mongodb-kubernetes-operator/pkg/authentication/authtypes"
 	"github.com/mongodb/mongodb-kubernetes-operator/pkg/automationconfig"
+	"github.com/mongodb/mongodb-kubernetes-operator/pkg/util/constants"
 	"github.com/stretchr/testify/assert"
 )
